Context is overkill for the form state, define it in a hook function in the wizard file
